+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [1.99.0] - Not released
+### Fixed
+- The gone user's comment likes was counted in some API responses and realtime
+  messages.
 
 ## [1.98.0] - 2021-04-08